Replaced solo Python files with project files
I realized I was being a little too paranoid about people using project metadata against me. So I decided to copy all the project files to the repo instead of just the Python ones. I also reasoned that it'd be more useful and easier to handle if others were able to see the processes that went into the creation and management of the projects.